move the "config"s from app.js and router.js to a new config.js file
